third_party/WebKit/LayoutTests/fast/dom/document-contentType-data-uri-expected.txt

PASS "text/xml" is "text/xml"
PASS "application/xml" is "application/xml"
PASS "image/svg+xml" is "image/svg+xml"
PASS "text/html" is "text/html"
PASS successfullyParsed is true

TEST COMPLETE


   
